feat: 기물에서 책임을 다하기 위해 BoardSnapShot 구현
