+++ src/main/java/chess/domain/board/Board.java
@@ -2,7 +2,6 @@ package chess.domain.board;
 
 import chess.domain.piece.Piece;
 import java.util.Collections;
-import java.util.List;
 import java.util.Map;
 import java.util.Optional;
 
@@ -18,53 +17,12 @@ public class Board {
         return Optional.ofNullable(board.get(square));
     }
 
-    public boolean canMove(final Square source, final List<Square> routes) {
-        final boolean isExistHurdle = isExistHurdle(routes.subList(0, routes.size() - 1));
-        final Square destination = routes.get(routes.size() - 1);
-        if (!board.containsKey(destination)) {
-            return !isExistHurdle;
-        }
-        return isAttackable(source, destination);
-    }
-
-    private boolean isExistHurdle(final List<Square> squares) {
-        return squares.stream()
-                .anyMatch(board::containsKey);
-    }
-
-    private boolean isAttackable(final Square source, final Square destination) {
-        return board.containsKey(destination) && !isSameColor(source, destination);
-    }
-
-    private boolean isSameColor(final Square source, final Square destination) {
-        final Piece sourcePiece = board.get(source);
-        final Piece destinationPiece = board.get(destination);
-        return sourcePiece.hasSameColor(destinationPiece);
-    }
-
-    public boolean canMovePawn(final Square source, final List<Square> routes) {
-        if (isDiagonalUnit(source, routes.get(0))) {
-            return isAttackable(source, routes.get(0));
-        }
-        return canMoveStraight(source, routes);
-    }
-
-    private boolean isDiagonalUnit(final Square source, final Square destination) {
-        final int distanceFile = destination.calculateDistanceFile(source);
-        final int distanceRank = destination.calculateDistanceRank(source);
-        return Math.abs(distanceFile) == Math.abs(distanceRank) && Math.abs(distanceFile) == 1;
-    }
-
-    private boolean canMoveStraight(final Square source, final List<Square> routes) {
-        final Piece piece = board.get(source);
-        final Square destination = routes.get(0);
-        final int distanceRank = Math.abs(destination.calculateDistanceRank(source));
-        return (distanceRank == 2 && source.isInitPawnPosition(piece.isBlack()))
-                || (distanceRank == 1 && !board.containsKey(destination));
-    }
-
     public void move(final Square source, final Square destination) {
         board.put(destination, board.remove(source));
+    }
+
+    public BoardSnapShot getBoardSnapShot() {
+        return BoardSnapShot.from(board);
     }
 
     public Map<Square, Piece> getBoard() {
