+++ src/main/java/chess/domain/piece/Pawn.java
@@ -1,7 +1,10 @@
 package chess.domain.piece;
 
+import chess.domain.board.BoardSnapShot;
+import chess.domain.board.Square;
 import chess.domain.piece.strategy.DirectStrategy;
 import chess.domain.piece.strategy.vector.DirectVector;
+import java.util.List;
 
 public class Pawn extends Piece {
 
@@ -10,7 +13,7 @@ public class Pawn extends Piece {
     }
 
     @Override
-    public boolean isPawn() {
-        return true;
+    public boolean isMovable(final Square source, final List<Square> route, final BoardSnapShot boardSnapShot) {
+        return boardSnapShot.canMovePawn(source, route);
     }
 }
