+++ src/main/java/chess/domain/piece/Rook.java
@@ -1,7 +1,10 @@
 package chess.domain.piece;
 
+import chess.domain.board.BoardSnapShot;
+import chess.domain.board.Square;
 import chess.domain.piece.strategy.SlidingStrategy;
 import chess.domain.piece.strategy.vector.SlidingVector;
+import java.util.List;
 
 public class Rook extends Piece {
 
@@ -10,7 +13,7 @@ public class Rook extends Piece {
     }
 
     @Override
-    public boolean isPawn() {
-        return false;
+    public boolean isMovable(final Square source, final List<Square> route, final BoardSnapShot boardSnapShot) {
+        return boardSnapShot.canMove(source, route);
     }
 }
