+++ src/test/java/chess/domain/board/BoardTest.java
@@ -7,6 +7,7 @@ import static chess.util.SquareFixture.A_TWO;
 import static chess.util.SquareFixture.B_EIGHT;
 import static chess.util.SquareFixture.B_ONE;
 import static chess.util.SquareFixture.B_SEVEN;
+import static chess.util.SquareFixture.B_THREE;
 import static chess.util.SquareFixture.B_TWO;
 import static chess.util.SquareFixture.C_EIGHT;
 import static chess.util.SquareFixture.C_ONE;
@@ -41,10 +42,8 @@ import chess.domain.piece.Knight;
 import chess.domain.piece.Pawn;
 import chess.domain.piece.Queen;
 import chess.domain.piece.Rook;
-import java.util.List;
 import org.junit.jupiter.api.DisplayNameGeneration;
 import org.junit.jupiter.api.DisplayNameGenerator;
-import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.api.Test;
 
 @SuppressWarnings("NonAsciiCharacters")
@@ -98,125 +97,12 @@ class BoardTest {
         );
     }
 
-    @Nested
-    class canMove_메서드는 {
-
-        @Nested
-        class 나이트의_경우 {
-            @Test
-            void 이동경로로_이동할_수_있으면_true_반환한다() {
-                final Board board = BoardFactory.create();
-                final Square source = new Square(File.B, Rank.ONE);
-                final List<Square> routes = List.of(new Square(File.C, Rank.THREE));
-
-                assertThat(board.canMove(source, routes)).isTrue();
-            }
-
-            @Test
-            void 이동경로로_이동할_수_없으면_false_반환한다() {
-                final Board board = BoardFactory.create();
-                final Square source = new Square(File.G, Rank.ONE);
-                final List<Square> routes = List.of(new Square(File.E, Rank.TWO));
-
-                assertThat(board.canMove(source, routes)).isFalse();
-            }
-        }
-
-        @Nested
-        class 룩의_경우 {
-
-            @Test
-            void 이동경로로_이동할_수_없으면_false_반환한다() {
-                final Board board = BoardFactory.create();
-                final Square source = new Square(File.A, Rank.ONE);
-                final List<Square> routes = List.of(new Square(File.A, Rank.TWO), new Square(File.A, Rank.THREE));
-
-                assertThat(board.canMove(source, routes)).isFalse();
-            }
-        }
-
-        @Nested
-        class 비숍의_경우 {
-
-            @Test
-            void 이동경로로_이동할_수_없으면_false_반환한다() {
-                final Board board = BoardFactory.create();
-                final Square source = new Square(File.C, Rank.ONE);
-                final List<Square> routes = List.of(new Square(File.D, Rank.TWO), new Square(File.E, Rank.THREE));
-
-                assertThat(board.canMove(source, routes)).isFalse();
-            }
-        }
-
-        @Nested
-        class 퀸의_경우 {
-
-            @Test
-            void 이동경로로_이동할_수_없으면_false_반환한다() {
-                final Board board = BoardFactory.create();
-                final Square source = new Square(File.D, Rank.ONE);
-                final List<Square> routes = List.of(new Square(File.C, Rank.TWO), new Square(File.B, Rank.THREE));
-
-                assertThat(board.canMove(source, routes)).isFalse();
-            }
-        }
-
-        @Nested
-        class 킹의_경우 {
-
-            @Test
-            void 이동경로로_이동할_수_없으면_false_반환한다() {
-                final Board board = BoardFactory.create();
-                final Square source = new Square(File.E, Rank.ONE);
-                final List<Square> routes = List.of(new Square(File.F, Rank.TWO));
-
-                assertThat(board.canMove(source, routes)).isFalse();
-            }
-        }
-    }
-
-    @Nested
-    class canMovePawn_메서드는 {
-
-        @Test
-        void 폰이_초기_위치일_때_두칸_갈_수_있다() {
-            final Board board = BoardFactory.create();
-            final Square source = new Square(File.A, Rank.TWO);
-            final List<Square> routes = List.of(new Square(File.A, Rank.FOUR));
-
-            assertThat(board.canMovePawn(source, routes)).isTrue();
-        }
-
-        @Test
-        void 폰이_초기_위치일_때_한칸_갈_수_있다() {
-            final Board board = BoardFactory.create();
-            final Square source = new Square(File.B, Rank.TWO);
-            final List<Square> routes = List.of(new Square(File.B, Rank.THREE));
-
-            assertThat(board.canMovePawn(source, routes)).isTrue();
-        }
-
-        @Test
-        void 폰이_초기_위치일_때_대각선으로_갈_수_없다() {
-            final Board board = BoardFactory.create();
-            final Square source = new Square(File.C, Rank.TWO);
-            final List<Square> routes = List.of(new Square(File.D, Rank.THREE));
-
-            assertThat(board.canMovePawn(source, routes)).isFalse();
-        }
-    }
-
     @Test
     void 기물을_움직인다() {
         final Board board = BoardFactory.create();
-        final Square source = new Square(File.E, Rank.TWO);
-        final Square destination = new Square(File.E, Rank.FOUR);
 
-        board.move(source, destination);
+        board.move(B_TWO, B_THREE);
 
-        assertAll(
-                () -> assertThat(board.findPieceOf(source)).isEmpty(),
-                () -> assertThat(board.findPieceOf(destination)).containsInstanceOf(Pawn.class)
-        );
+        assertThat(board.findPieceOf(B_THREE)).containsInstanceOf(Pawn.class);
     }
 }
