+++ src/test/java/chess/domain/piece/BishopTest.java
@@ -36,5 +36,4 @@ class BishopTest {
                 .isThrownBy(() -> rook.findRoute(H_EIGHT, D_FIVE))
                 .withMessage("해당 기물이 움직일 수 있는 경로가 아닙니다.");
     }
-
 }
